+++ src/Test/Scene/physicsTestScene.ts
@@ -13,11 +13,13 @@ import { Vector3 } from "@babylonjs/core/Maths/math.vector";
 // import { CreateBox } from "@babylonjs/core/Meshes/Builders/boxBuilder";
 import type { Mesh } from "@babylonjs/core/Meshes/mesh";
 import { TransformNode } from "@babylonjs/core/Meshes/transformNode";
+// import { HavokPlugin } from "@babylonjs/core/Physics/v2/Plugins/havokPlugin";
 // import { PhysicsImpostor } from "@babylonjs/core/Physics/v1/physicsImpostor";
 import { DepthOfFieldEffectBlurLevel } from "@babylonjs/core/PostProcesses/depthOfFieldEffect";
 import { DefaultRenderingPipeline } from "@babylonjs/core/PostProcesses/RenderPipeline/Pipelines/defaultRenderingPipeline";
 import { Scene } from "@babylonjs/core/scene";
 
+// import havok from "@babylonjs/havok";
 import type { MmdAnimation } from "@/Loader/Animation/mmdAnimation";
 import type { MmdStandardMaterialBuilder } from "@/Loader/mmdStandardMaterialBuilder";
 import type { BpmxLoader } from "@/Loader/Optimized/bpmxLoader";
@@ -29,6 +31,7 @@ import type { MmdMesh } from "@/Runtime/mmdMesh";
 import { MmdRuntime } from "@/Runtime/mmdRuntime";
 import { MmdAmmoJSPlugin } from "@/Runtime/Physics/mmdAmmoJSPlugin";
 import { MmdAmmoPhysics } from "@/Runtime/Physics/mmdAmmoPhysics";
+// import { MmdPhysics } from "@/Runtime/Physics/mmdPhysics";
 import { MmdPlayerControl } from "@/Runtime/Util/mmdPlayerControl";
 
 import type { ISceneBuilder } from "../baseRuntime";
@@ -108,9 +111,9 @@ export class SceneBuilder implements ISceneBuilder {
             }],
             ["physics", async(updateProgress): Promise<void> => {
                 updateProgress({ lengthComputable: true, loaded: 0, total: 1 });
-                const ammoInstance = await ammo();
-                const ammoPlugin = new MmdAmmoJSPlugin(true, ammoInstance);
-                scene.enablePhysics(new Vector3(0, -9.8 * 10, 0), ammoPlugin);
+                const physicsInstance = await ammo();
+                const physicsPlugin = new MmdAmmoJSPlugin(true, physicsInstance);
+                scene.enablePhysics(new Vector3(0, -9.8 * 10, 0), physicsPlugin);
                 updateProgress({ lengthComputable: true, loaded: 1, total: 1 });
             }]
         ]);